+++ subscribers_db.sql
@@ -11,14 +11,19 @@ CREATE TABLE status (
                    NOT NULL,
     IPV6 DECIMAL (32) UNIQUE
                    NOT NULL
-
+    IPV4_TEXT TEXT (512) NOT NULL,
+    IPV6_TEXT TEXT (512) NOT NULL
 );
 
 INSERT INTO status (
                        IPV4,
-                       IPV6
+                       IPV6,
+		       IPV4_TEXT,
+                       IPV6_TEXT
                    )
                    VALUES (
+                       1,
+                       1,
                        1,
                        1
                    );